feat(competition): add InscriptionsInfos component and update Overview layout
